requirements.txt
python-dotenv
spotipy
scikit-learn
transformers
pandas
numpy